backend/recommendations/algorithms/clustering.py
import pandas as pd
from sklearn.cluster import KMeans
from sklearn.preprocessing import StandardScaler

def cluster_users(user_data):
    scaler = StandardScaler()
    X_scaled = scaler.fit_transform(user_data)

    kmeans = KMeans(n_clusters=3, random_state=42)
    clusters = kmeans.fit_predict(X_scaled)

    return clusters
