requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
scikit-learn
surprise
lightfm
